Updating to current version
Lots of changes:

Finalised everything on loading screen, instructions screen (except
actual text content), questions screen (except icons for
questions/trivia/tasks/camera), and changes to helper classes to allow
for camera and saving functionality.

Adjusted the way in which the background images are displayed as well,
and added padding to accommodate for the iOS status bar on every screen
I've done.